Move external subcommand to external mod
