+++ crates/flowctl/src/main.rs
@@ -2,7 +2,7 @@ mod combine;
 mod external;
 
 use clap::Parser;
-use external::{exec_go_flowctl, ExternalArgs};
+use external::{exec_go_flowctl, ExternalSubcommand};
 
 /// flowctl is a CLI for interacting with Flow data planes (and soon, control planes).
 #[derive(Debug, Parser)]
@@ -20,53 +20,6 @@ enum Subcommand {
     External(ExternalSubcommand),
 }
 
-/// External subcommands are those that are provided by the flowctl Go-based binary.
-#[derive(Debug, clap::Subcommand)]
-#[clap(rename_all = "kebab-case")]
-enum ExternalSubcommand {
-    /// Low-level APIs for automation
-    Api(ExternalArgs),
-    /// Check a Flow catalog for errors
-    Check(ExternalArgs),
-    /// Build a catalog and deploy it to a data plane
-    Deploy(ExternalArgs),
-    /// Discover available captures of an endpoint
-    Discover(ExternalArgs),
-    /// Interact with broker journals
-    Journals(ExternalArgs),
-    /// Print the catalog JSON schema
-    JsonSchema(ExternalArgs),
-    /// Print combined configuration and exit
-    PrintConfig(ExternalArgs),
-    /// Serve a component of Flow
-    Serve(ExternalArgs),
-    /// Interact with consumer shards
-    Shards(ExternalArgs),
-    /// Run an ephemeral, temporary local data plane
-    TempDataPlane(ExternalArgs),
-    /// Locally test a Flow catalog
-    Test(ExternalArgs),
-}
-
-impl ExternalSubcommand {
-    fn into_subcommand_and_args(self) -> (&'static str, ExternalArgs) {
-        use ExternalSubcommand::*;
-        match self {
-            Api(a) => ("api", a),
-            Check(a) => ("check", a),
-            Deploy(a) => ("deploy", a),
-            Discover(a) => ("discover", a),
-            Journals(a) => ("journals", a),
-            JsonSchema(a) => ("json-schema", a),
-            PrintConfig(a) => ("print-config", a),
-            Serve(a) => ("serve", a),
-            Shards(a) => ("shards", a),
-            TempDataPlane(a) => ("temp-data-plane", a),
-            Test(a) => ("test", a),
-        }
-    }
-}
-
 #[derive(Debug, clap::Subcommand)]
 #[clap(rename_all = "kebab-case")]
 enum InternalSubcommand {
@@ -79,8 +32,11 @@ fn main() -> Result<(), anyhow::Error> {
     // top-level arguments. If it does, then it will exit(0) automatically. This will not be the
     // case for external subcommands, though, as they handle their own --help and --version flags.
     let cli = Cli::parse();
+    run_subcommand(cli.subcommand)
+}
 
-    match cli.subcommand {
+fn run_subcommand(subcommand: Subcommand) -> Result<(), anyhow::Error> {
+    match subcommand {
         Subcommand::Internal(internal) => execute_internal_subcommand(internal),
         Subcommand::External(external) => {
             let (subcommand, external_args) = external.into_subcommand_and_args();
